requirements.txt
streamlit>=1.36.0
openai>=1.40.0
sentence-transformers>=3.0.1
chromadb>=0.5.5
python-dotenv>=1.0.1
